cmd/gomobile: drop init step for build and bind commands
With the new Go 1.10 build cache, the prebuild work by the gomobile
init command has become useless. Use the opportunity to simplify
init to only building OpenAL for Android, and to drop the requirement
to run gomobile init before gomobile build and gomobile bind.

Add the -ndk flag to gomobile init and gomobile bind to specify
the path to the NDK if it can't be auto-detected from ANDROID_HOME.

Change-Id: I4692fcaa927e7076a6387d080ebc1726905afd72